requirements.txt
pandas
python-dotenv
openai
pinecone
langchain-text-splitters
